+++ tests/stack_test.c
@@ -14,7 +14,7 @@ typedef struct queue_elt {
 }queue_elt;
 
 static int values[] = { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 };
-#define VALUES_MAX sizeof_array(values)
+#define VALUES_MAX (int)sizeof_array(values)
 
 static void
 basic_tests(struct zt_unit_test *test, void *data UNUSED)
